Méthodes testEmployePlusRiche correctes. Probleme de données d'initialisation.
