+++ src/test/java/com/ipiecoles/java/java350/repository/EmployeRepositoryTest.java
@@ -20,9 +20,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 import java.util.HashSet;
 import java.util.List;
 
-/**
- * Created by HCHARBONNEYR on 12/03/2018.
- */
+
 @RunWith(SpringRunner.class)
 //@DataJpaTest
 @SpringBootTest(classes = MySpringApplication.class)
@@ -155,25 +153,6 @@ public class EmployeRepositoryTest {
         Assertions.assertThat(listeEmploye).isNotEmpty();
         Assertions.assertThat(listeEmploye).hasSize(1);
         Assertions.assertThat(listeEmploye).contains(m);
-    }
-
-    @Test
-    public void testFindEmployePlusRicheSalaireIdentique()
-    {
-        //GIVEN
-        Technicien t = new Technicien("novet",  "cyrille",  "CN11", null ,  2000d,1 );
-        employeRepository.save(t);
-        Manager m = new Manager("moufid",  "manal",  "AL123", null ,  2000d, new HashSet<>());
-        employeRepository.save(m);
-        Commercial c=new Commercial("dubois","sandrine","DS345",null,2000d,2345d);
-        employeRepository.save(c);
-
-        //WHEN
-        List<Employe> listeEmploye =employeRepository.findEmployePlusRiches();
-
-        //THEN
-        Assertions.assertThat(listeEmploye).isEmpty();
-
     }
